+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `Client.iter_commands`, `Client.iter_message_commands` and `Client.iter_slash_commands`.
+
+### Removed
+- `tanjun.abc.ExecutableCommand.execute` and `check_context` as this doesn't work typing wise.
+  This doesn't effect the implementations nor full command types.
 
 ## [2.1.1a1] - 2021-10-09
 ### Added